Move utility function in api_camps_routes to services. Fix linting Replace vars with const and let.
