+++ api/services/index.js
@@ -1,0 +1,9 @@
+const helperService = require('./helper.service'),
+    campsService = require('./camps.service'),
+    usersService = require('./users.service');
+
+module.exports = {
+    helperService,
+    campsService,
+    usersService
+};
